Conflicts:
	config/arm64/default
	config/armv5tel/versatile
	config/armv7hl/cubox
	config/armv7hl/default
	config/armv7hl/exynos
	config/armv7hl/imx51
	config/armv7hl/lpae
	config/armv7hl/u8500
	config/i386/debug
	config/i386/default
	config/i386/desktop
	config/i386/ec2
	config/i386/pae
	config/i386/trace
	config/i386/vanilla
	config/i386/xen
	config/ia64/debug
	config/ia64/default
	config/ia64/trace
	config/ia64/vanilla
	config/ppc/default
	config/ppc/vanilla
	config/ppc64/debug
	config/ppc64/default
	config/ppc64/trace
	config/ppc64/vanilla
	config/s390/s390
	config/s390/trace
	config/s390/vanilla
	config/s390x/default
	config/s390x/trace
	config/s390x/vanilla
	config/x86_64/debug
	config/x86_64/default
	config/x86_64/desktop
	config/x86_64/ec2
	config/x86_64/trace
	config/x86_64/vanilla
	config/x86_64/xen
	patches.xen/pci-guestdev
	patches.xen/xen-x86-EFI
	patches.xen/xen3-auto-xen-kconfig.diff
	patches.xen/xen3-fixup-xen
	patches.xen/xen3-patch-3.1
	patches.xen/xen3-patch-3.7
	series.conf

suse-commit: c967b183af3f43a2c8eeece2780aa4c077b0139c

+++ drivers/xen/blkback/common.h
@@ -85,15 +85,15 @@ typedef struct blkif_st {
 
 	/* statistics */
 	unsigned long       st_print;
-	int                 st_rd_req;
-	int                 st_wr_req;
-	int                 st_oo_req;
-	int                 st_br_req;
-	int                 st_fl_req;
-	int                 st_ds_req;
-	int                 st_pk_req;
-	int                 st_rd_sect;
-	int                 st_wr_sect;
+	unsigned long       st_rd_req;
+	unsigned long       st_wr_req;
+	unsigned long       st_oo_req;
+	unsigned long       st_br_req;
+	unsigned long       st_fl_req;
+	unsigned long       st_ds_req;
+	unsigned long       st_pk_req;
+	unsigned long       st_rd_sect;
+	unsigned long       st_wr_sect;
 
 	wait_queue_head_t waiting_to_free;
 } blkif_t;
